pyproject.toml

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "expy"
version = "0.0.1"
description = "A minimal AI experimentation framework"
readme = "README.md"
authors = [
  { name = "Kerry Cerqueira", email = "kerrycerqueira@gmail.com" }
]
dependencies = [
	"pyserde>=0.23",
	"nbclient>=0.10.2",
	"typer>=0.9.0"
]
requires-python = ">=3.12"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
expy = "expy.cli:expy"
